update cosineSinilar api

+++ DTOs/Auth/FaceLoginRequestDto.cs
@@ -3,6 +3,7 @@ namespace MeetingBackend.DTOs.Auth;
 public class FaceLoginRequestDto
 {
     // Embedding vector từ thiết bị (cùng kích thước với embedding đã lưu trong DB).
-    public float[] Embedding { get; set; } = Array.Empty<float>();
+    // public byte[] Embedding { get; set; } = Array.Empty<byte>();
+    public int[] Embedding { get; set; } = Array.Empty<int>();
 }
